+++ internal/commands/all/all.go
@@ -8,6 +8,7 @@ import (
 	databaseindex "github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/database/index"
 	databaseproperties "github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/database/properties"
 	databasesession "github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/database/session"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/gateway"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/ipaddress"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/kubernetes"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/kubernetes/nodegroup"
@@ -188,6 +189,11 @@ func BuildCommands(rootCmd *cobra.Command, conf *config.Config) {
 	commands.BuildCommand(objectstorage.DeleteCommand(), objectStorageCommand.Cobra(), conf)
 	commands.BuildCommand(objectstorage.ListCommand(), objectStorageCommand.Cobra(), conf)
 	commands.BuildCommand(objectstorage.ShowCommand(), objectStorageCommand.Cobra(), conf)
+
+	// Network Gateway operations
+	gatewayCommand := commands.BuildCommand(gateway.BasegatewayCommand(), rootCmd, conf)
+	commands.BuildCommand(gateway.DeleteCommand(), gatewayCommand.Cobra(), conf)
+	commands.BuildCommand(gateway.ListCommand(), gatewayCommand.Cobra(), conf)
 
 	// Misc
 	commands.BuildCommand(