+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.4.0] - 2024-02-08
+
 ### Added
 
 - Add `gateway` commands (`delete`, `list`) for Network gateway management.
@@ -344,7 +346,8 @@ Initial public beta release :tada:
 ### Added
 - Current feature set added! First internal release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.3.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.4.0...HEAD
+[3.4.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.2.2...v3.3.0
 [3.2.2]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.2.1...v3.2.2
 [3.2.1]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.2.0...v3.2.1